+++ CHANGELOG.md
@@ -65,3 +65,4 @@ instance you've created and share it around with friends to simply import into t
 - Add host to ip resolution to Network Checker Tool
 - Allow setting network connection timeouts
 - Remove pack tags and add pack for semi public packs
+- Fix main Forge library missing caused packs not able to redownload it